backend/requirements.txt
alembic==1.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-Restless==0.17.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
h11==0.8.1
h2==3.1.1
hpack==3.0.0
hstspreload==2019.11.4
httpx==0.7.6
hyperframe==5.2.0
idna==2.8
itsdangerous==1.1.0
jellyfish==0.5.6
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.2
marshmallow-sqlalchemy==0.19.0
mimerender==0.6.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-mimeparse==1.6.0
python-status==1.0.1
requests==2.22.0
rfc3986==1.3.2
simple-rest-client==1.0.0
six==1.13.0
SQLAlchemy==1.3.10
urllib3==1.25.6
us==1.0.0
Werkzeug==0.16.0
